+++ controllers/forgotPasswordController.js
